% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allranks.R
\docType{data}
\name{allranks}
\alias{allranks}
\title{allranks}
\format{
An object of class \code{data.frame} with 2519 rows and 9 columns.
}
\usage{
allranks
}
\description{
'allranks' containing all information of performance assessment, criteria selected and ranking.
}
\keyword{datasets}
