% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EIGENMS.R
\name{EIGENMS}
\alias{EIGENMS}
\title{EigenMS normalization}
\usage{
EIGENMS(data, label)
}
\arguments{
\item{data}{This is the first ones argument}

\item{label}{This is the second ones argument}
}
\description{
EigenMS estimates and preserves fixed effects.
Random effects may be attempted in the future.
}
