% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqvisualize.R
\name{lfqvisualize}
\alias{lfqvisualize}
\title{lfqvisualize}
\usage{
lfqvisualize(object, top = 100)
}
\arguments{
\item{object}{The input is the output file of the "lfq_access" or "lfq_spiked".}

\item{top}{The default 'top' value is 100. You can view the top ranking heatmap you want.}
}
\description{
lfqvisualize
}
\details{
Draw heatmap and save as ProteoLFQ-OUTPUT.Figure-Top.XXX.workflows.pdf. For function definitions and descriptions please use "??ProteoLFQ" command in R.
}
\examples{
library(EVALFQ)
\donttest{data_q <- PrepareInuputFiles(acquisitionmethods=2,
rawdataset = "MaxQuant_proteinGroups_LFQ.txt", lable = "MaxQuant_LFQ_Label.txt")}
\donttest{lfqevalueall(data_q = data_q,
assum_a="Y", assum_b="Y", assum_c="Y", Ca="1", Cb="1", Cc="1", Cd="1")}
\donttest{lfqvisualize(object = "EVALFQ-OUTPUT.Data-Overall.Ranking.csv", top = 100)}
}
