% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_df.R
\docType{data}
\name{my_df}
\alias{my_df}
\title{my_df}
\format{
An object of class \code{data.frame} with 28 rows and 500 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/29276171}
}
\usage{
my_df
}
\description{
The dataset PXD006129 included 14 western-style diet mice vs 14 chow-fed mice. (Cell Host Microbe, 23:27–40, 2018)
}
\keyword{datasets}
