% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareInuputFiles.R
\name{PrepareInuputFiles}
\alias{PrepareInuputFiles}
\title{Prepare input file in standard formats of quantification tools}
\usage{
PrepareInuputFiles(acquisitionmethods, rawdataset, lable)
}
\arguments{
\item{acquisitionmethods}{Input the corresponding "number" of acquisition techniques as follows:
If set 1, the user chooses to process the data based on SWATH-MS.
If set 2, the user chooses to process the data based on Peak Intensity.
If set 3, the user chooses to process the data based on Spectral Counting.}

\item{rawdataset}{Input the name of your raw dataset directly obtained from software.}

\item{lable}{Input the label of your dataset.}
}
\value{
prepareinputfile matrix
}
\description{
process the output files obtained from popular proteomic quantification software (>=18).
}
\examples{
library(EVALFQ)
\donttest{data_s <- PrepareInuputFiles(acquisitionmethods=2,
rawdataset = "MaxQuant_proteinGroups_LFQ.txt", lable = "MaxQuant_LFQ_Label.txt")}
}
