% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectworkflows.R
\docType{data}
\name{selectworkflows}
\alias{selectworkflows}
\title{selectworkflows}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/29276171}
}
\usage{
selectworkflows
}
\description{
Input the name of your prefered strategies.
}
\keyword{datasets}
