% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiked_data.R
\docType{data}
\name{spiked_data}
\alias{spiked_data}
\title{spiked_data}
\format{
An object of class \code{data.frame} with 6 rows and 14 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/26321463}
}
\usage{
spiked_data
}
\description{
The dataset PXD002099 included 48 UPS1 proteins spiked with distinct concentrations (10 vs 25 fmol/ul) (J. Proteome Res. 11:4118–4126, 2015).
}
\keyword{datasets}
